+++ package.json
@@ -2,8 +2,11 @@
   "name": "arkvatar-ts",
   "version": "1.0.0",
   "description": "Typescript package for interacting with the Arkvatar API.",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
   "scripts": {
     "test": "jest --config jestconfig.json",
     "build": "tsc",
