+++ tsconfig.json
@@ -3,7 +3,7 @@
     "target": "es2018",
     "module": "commonjs",
     "declaration": true,
-    "outDir": "./lib",
+    "outDir": "./dist",
     "strict": true
   },
   "include": ["src"],
